Combine Interface: SegmentSelector; libcpu: permissions parameter
